# Conflicts:
#	src/main/kotlin/com/tastypizza/menuorders/controllers/OrderController.kt
#	src/main/kotlin/com/tastypizza/menuorders/services/OrderService.kt

+++ src/main/kotlin/com/tastypizza/menuorders/entities/MenuItemOption.kt
@@ -10,7 +10,7 @@ import javax.persistence.*
 @Setter
 @Table(name="menuItemOption")
 data class MenuItemOption (
-    var count: Int=0,
+    var count: Long=0,
     val name: String="",
     val pictureId: String="",
     val price: Int=0,
